@ajwheeler 2024-07-18
# An update on performance

These times are for the radiative transfer calculation _only_, but the input absorption data came from these parameters:
```julia
linelist = [Korg.get_VALD_solar_linelist() ;  Korg.get_APOGEE_DR17_linelist()] 
atm = interpolate_marcs(5000.0, 3.0)
patm = Korg.PlanarAtmosphere(atm)
A_X = format_A_X()
wl_grid = [3000:0.01:5000, 16000:0.01:16500]
```

## old transfer implementation
This is for the default implementation (anchored + linear-flux-only in the new lingo).  I'm primarily concerned with the performance in the default case.
- planar (using $E_2$):  `282.709 ms (250013 allocations: 131.61 MiB)`
- spherical (using multiple rays)l: `1.541 s (4842296 allocations: 605.94 MiB)`

## new implementation
Again, I'm concerned with the defaults here, so I'm not evaluating bezier $\tau$ or $I$ integrals.
| | spherical | planar|
|-|-|-|
| **flux-only** | `1.778 s (6000348 allocations: 2.35 GiB)` |`292.917 ms (250040 allocations: 137.34 MiB)` |
| **$I$ retained**| `3.274 s (6000348 allocations: 4.81 GiB)` | `3.285 s (5000203 allocations: 4.66 GiB)` |

So there is a small slowdown for the default methods.  I'm resisting the urge to further optimize, though I will note that the increased allocations suggests to me that there's some low-hanging fruit here.